Change get_valid_date return type to string
get_valid_date is the function that sanitize the user input during
the upload process and returns the sanitized input to the rest of the
program. Before this commit, this function returned a pd.Timestamp
object. Since the date (as all the data) is ultimately kept as
string, so making sure this function returns a string, and not some
abstract object, allow us to control the string format directly. In
particular, it allow us to make sure that the string that save the
creation date is of the format "%Y-%m-%d" and not "%Y-%m-%d %H:%M:%S"
which is the default pd.Timestamp string format.