Make the programs read data from the bucket
Previously, the programs relied on the existence of the data locally,
and got the paths as arguments. With this change, the programs will get
the keynames of the files in bucket, and will read them from the bucket
directly, removing the dependency on the local state of the hard disk.
The commit also adds needed dependency to the experiment-management
package from our organizational GitHub.