Remove data integrity validation when rekeying an object
Rekeying is done by downloading a file, and uploading it again with a different key.
Previously, we made sure that the file downloaded correctly by comaparing its MD5 hash
with the ETag of the object. However, the ETag is not guaranteed to be the MD5 hash of the
file (even though it sometimes is) and therefore we cannot utilize it for this functionality.