Fix a crash for empty dataframes
If the forecast_data method receives an empty data (for example, if
there exists data in the current year, but not in past years), depending
on the chosen model, the program might crash. To alleviate this issue, we
introduce a new 'DummyModel' which is used only if the data is empty,
that forecasts zeroes for the requested time frame by the caller
function. The requested time frame is deduced by an additional argument
that is used only by this case, and is computed by the current year and
current month arguments of the program.