Fix wrong duplication scheme
Previously, due to oversight, the program would "deduplicate" the
invoices table on the preprocessing stage, by deleting all invoices
beyond the first invoice for each PO, PO item and fund year. We remove
this "deduplication" scheme, and instead deduplicate the orders table
(which is the only table that may have duplicated entries due to change
of budget attribute over the year), by summing over the PO-net-value of
all the duplicated entries, and keeping the earliest attributes. Since
only a very small portion of the POs suffer from this predicament, this
should not harm the model.